Merge Tool
Merge tool has been implemented
Works similar to Lock tool, enable it in the Tool Manager and use A/B on the right controller

Two things need to be done however before we can move on
- Add to editing space
- Fix out of bounds errors when updating edges

Line 296
The error comes from deleting one of the vertices when you merge two
Removing it from the vertices array creates an out of bounds error in UpdateMesh();, since we use vertex ids to access their indices

e.g. If we merge vertices 4 and 5, vertex 4 will be deleted, all edges from 4 are remapped to 5. If an edge connects vertices 4 and 7 (now 5 and 7), when we try to access vertices[7], we go out of bounds since vertices is only length 7, we're trying to access index 8